eruditefx_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
reportlab==4.2.2
Pillow==10.4.0
python-dateutil==2.9.0.post0
